pyproject.toml
[project]
name = "TLE-tracker"
version = "0.2.1"
description = ""
authors = [
    {name = "dawid.robak",email = "dawid.robak1704@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

[tool.poetry]
packages =[
    {include = "tle_tracker"},
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
poetry = "^2.1.2"

[tool.poetry.dependencies]
uvicorn = "^0.34.2"
skyfield = "^1.53"
paho-mqtt = "^2.1.0"
watchdog = "^6.0.0"
